+++ CHANGELOG.md
@@ -129,3 +129,4 @@
 - 2023-10-12: Metrics qualifier, scheduler, gp, session, +5
 - 2023-10-13: Retry gp, session, deals, scheduler, +2
 - 2023-10-13: Retry deals, session, gp, scheduler, +2
+- 2023-10-14: Pool session, gp, scheduler, qualifier, +2